refactor: extract DifferPort and wasm provider
